+++ libs/spark/src/Radio/Radio.stories.tsx
@@ -94,7 +94,7 @@ export const StatesHover: Story = PseudoStatesTemplate.bind({});
 StatesHover.parameters = { pseudo: { hover: true } };
 
 export const StatesFocus: Story = PseudoStatesTemplate.bind({});
-StatesFocus.parameters = { pseudo: { focus: true } };
+StatesFocus.parameters = { pseudo: { focusVisible: true } };
 
 const LabeledStatesTemplate = () => (
   <div style={{ display: 'flex', gap: '1rem', margin: '1rem' }}>
@@ -118,4 +118,4 @@ export const LabeledStatesHover: Story = PseudoLabeledStatesTemplate.bind({});
 LabeledStatesHover.parameters = { pseudo: { hover: true } };
 
 export const LabeledStatesFocus: Story = PseudoLabeledStatesTemplate.bind({});
-LabeledStatesFocus.parameters = { pseudo: { focus: true } };
+LabeledStatesFocus.parameters = { pseudo: { focusVisible: true } };